feat: dashboard dinamico, lista de cobros y correccion de qr token

+++ gym.api/src/main/java/com/snzalx/gym/api/dto/MembresiaDTO.java
@@ -1,13 +1,14 @@
 package com.snzalx.gym.api.dto;
 
 import lombok.Data;
+import java.math.BigDecimal; // Importar BigDecimal
 import java.time.LocalDate;
 import java.util.UUID;
 
 @Data
 public class MembresiaDTO {
     private UUID socioId;
-    private Double monto;
+    private BigDecimal monto; // Cambiado de Double a BigDecimal
     private String metodoPago;
     private Integer duracionDias; // Ejemplo: 30 para un mes
